+++ carveracontroller/addons/probing/ProbingPopup.kv
@@ -1,5 +1,5 @@
 
-#:include addons/probing/ProbingPreviewPopup.kv
+#:include addons/probing/preview/ProbingPreviewPopup.kv
 
 <ProbingPopup>:
     title:"Probing"
